Support form elements with no or duplicit id
This also improve compatibility with Grido as all inline-edit elements has same ID and previously it showed error message only for first one.

This commit changes way of determining the messageElement for holding error message. It holds id of the message element as data-attribute of the form element itself and uses always unique ID.