Fix 'filled' validation of inputs array (BC break)
Previously Nette.getValue() return array of values, even if that value inside the array was empty. So for 2 empty text inputs you would get ['', ''] which is then validated as "not empty" because it is array of values. This commit changes this behavior and doesn't add empty strings to this array which results in correct validation of 'filled' rule.

I'm not sure if this breaks any other existing use-case, but it's potential BC break.