Improve generating message ID
This improves generated IDs of message element to be nicer - e.g. when element is missing id, we use it's name as base for new id.
Also we make sure to not create multiple message elements for multi inputs (with same name) and generated ID we save to first input only.